Expose the response body outside
There's a few changes here. First, sendRequest has been renamed to Do,
so that the method is exported.

This make it possible to write very simple prototypes in a playground
such as:

	client := dnsimple.NewClient(apiToken, email)
	client.Do("GET", "domains/weppos.net/zone", nil, nil)

Second, the response body is now written to v, if the type implementa
io.Writer. This is a neat idea borrowed from go-github. Combined with
previous change, you can write even easier prototypes

	type Debugger struct {
	}

	func (d *Debugger) Write(p []byte) (n int, err error) {
		fmt.Println(string(p[:]))
		return 0, nil
	}

	func main() {
		client := dnsimple.NewClient("...", "...")
		client.Do("GET", "domains/weppos.net/zone", nil, &Debugger{})
	}