Replace custom "url" serializer with go-querystring
I think we hit the limit of our current implementation. We currently
don't handle nested structures, whereas go-querystring provides the
"same" feature but it also does.

It was a nice game to create such custom `url:` handler, but it is
worthless to make further changes as I end up reimplementing pretty
much the same amount of code.