Fix zone record tests
I also split the region tests from the main test, and copied the same
specs to both create and update.

I prefer to have duplicates tests, than testing different cases in
different methods (e.g. "global" in update and array in created,
because it is very fragile. if we refactor a single endpoint we risk to
lose half the tests).