Change client initialization to accept an http.Client
This implementation follows the most common best practices around using a RoundTripper for authentication, rather than passing a credential and modifying the Do to apply the credential headers.

Definining a RoundTripper is a more flexible, powerful, and elengant approach.

This change requires to bump to 1.7 (given we add the dependency to golang/x/oauth2 that uses the context package).

Closes GH-15