Export returned values (#91)
In #54 I decided to change the method signatures to not export the response values. At a distance of few years, I believe this was probably not a good idea.

While it works in most cases, it makes more complex to implement testing. An example is when you need to create tests that will construct pagination.

https://github.com/kubernetes-sigs/external-dns/blob/master/provider/dnsimple_test.go#L105-L128

golint is also quite vocal about how _it may be annoying_ to not export the values.

I'm rolling back the changes in this PR. The good news is that since we forced the callers to never reference the value directly, this change should have very limited impact on anyone using the client.