Handle case where ZoneRecord name can be not present (#92)
A ZoneRecord name can be an empty value `""`. This is because our API mimic the UI and it requires the consumer to specify the zone record name without the zone, in other words not the qualified name.

As a result, when the qualified name also matches the zone name (for apex), the value of name is blank:

```
zone: example
record: www.example.com
-> name: "www"

zone: example
record: example.com
-> name: ""
```

In Go, there is no way to distinguish a non-present string (null) vs a blank string. That's because [the zero value](https://tour.golang.org/basics/12) of a string type is an empty string.

I decided to not change the signature of ZoneRecord but instead supply a different type of record attributes. This is actually another quite common pattern for particularly complex requests, and we also use in the [Registrar](https://github.com/dnsimple/dnsimple-go/blob/master/dnsimple/registrar.go#L134) where we supply special input structs to handle the payload of the registrar calls. This time I went with `-Attributes` and not `-Input` as there is a direct relationship between this struct at the zone record attributes, as well the zone record struct.

Fixes https://github.com/dnsimple/dnsimple-go/issues/33